requirements.txt
langchain
numpy
streamlit
langchain-community
PyPDF2
unstructured
nltk
sentence-transformers
langchain-huggingface
faiss-cpu
python-docx
beautifulsoup4
requests
tqdm
langchain
faiss-cpu
transformers
sentence-transformers
torch
